refactor: rename inner package datamaxi/datamaxi -> datamaxi/resources
The doubled 'datamaxi.datamaxi' package name read like an accident and
leaked into tracebacks / __module__. Rename the inner package to
'resources' and update all internal imports, tests, and the docs
mkdocstrings ':::' directives.

Public surface unchanged: 'from datamaxi import Datamaxi' still works.
setuptools auto-discovery picks up datamaxi.resources (verified via
find_packages). Deep imports of 'datamaxi.datamaxi.*' (essentially
internal) must move to 'datamaxi.resources.*'.
