+++ datamaxi/__init__.py
@@ -1,4 +1,4 @@
-from datamaxi.datamaxi import Datamaxi  # noqa: F401
+from datamaxi.resources import Datamaxi  # noqa: F401
 from datamaxi.telegram import Telegram  # noqa: F401
 from datamaxi.naver import Naver  # noqa: F401
 from datamaxi.lib.constants import (  # noqa: F401
