+++ datamaxi/resources/cex_candle.py
@@ -3,7 +3,7 @@ import pandas as pd
 from datamaxi.api import Resource
 from datamaxi.lib.utils import check_required_parameter
 from datamaxi.lib.utils import check_required_parameters
-from datamaxi.datamaxi.utils import convert_data_to_data_frame
+from datamaxi.resources.utils import convert_data_to_data_frame
 from datamaxi.lib.constants import SPOT, FUTURES, INTERVAL_1D, USD, Market, Interval
