+++ datamaxi/resources/cex_symbol.py
