+++ datamaxi/resources/forex.py
