+++ tests/test_cex_ticker.py
@@ -6,7 +6,7 @@ import pandas as pd
 import pytest
 from urllib.parse import urlparse, parse_qs
 
-from datamaxi.datamaxi.cex_ticker import CexTicker
+from datamaxi.resources.cex_ticker import CexTicker
 from datamaxi.error import ClientError, ServerError
 from tests.util import mock_http_response
