+++ tests/test_premium.py
@@ -6,7 +6,7 @@ import pandas as pd
 import pytest
 from urllib.parse import urlparse, parse_qs
 
-from datamaxi.datamaxi.premium import Premium
+from datamaxi.resources.premium import Premium
 from datamaxi.error import ClientError, ServerError
 from tests.util import mock_http_response
