+++ tests/test_query_params.py
@@ -11,8 +11,8 @@ import re
 import responses
 from urllib.parse import urlparse, parse_qs
 
-from datamaxi.datamaxi.liquidation import Liquidation
-from datamaxi.datamaxi.open_interest import OpenInterest
+from datamaxi.resources.liquidation import Liquidation
+from datamaxi.resources.open_interest import OpenInterest
 
 BASE_URL = "https://api.datamaxiplus.com"
